+++ scripts/detection/faster_rcnn/train_faster_rcnn.py
@@ -11,27 +11,15 @@ os.environ['MXNET_EXEC_BULK_EXEC_MAX_NODE_TRAIN_BWD'] = '25'
 os.environ['MXNET_GPU_COPY_NTHREADS'] = '1'
 os.environ['MXNET_OPTIMIZER_AGGREGATION_SIZE'] = '54'
 
-import logging
-import time
-import numpy as np
-import mxnet as mx
-from mxnet import gluon
 from mxnet.contrib import amp
 import gluoncv as gcv
 
 gcv.utils.check_version('0.7.0')
 from gluoncv import data as gdata
 from gluoncv import utils as gutils
-from gluoncv.model_zoo import get_model
-from gluoncv.data.batchify import FasterRCNNTrainBatchify, Tuple, Append
-from gluoncv.data.transforms.presets.rcnn import FasterRCNNDefaultTrainTransform, \
-    FasterRCNNDefaultValTransform
+from gluoncv.estimators.rcnn.faster_rcnn import FasterRCNNEstimator
 from gluoncv.utils.metrics.voc_detection import VOC07MApMetric
 from gluoncv.utils.metrics.coco_detection import COCODetectionMetric
-from gluoncv.utils.parallel import Parallel
-from gluoncv.utils.metrics.rcnn import RPNAccMetric, RPNL1LossMetric, RCNNAccMetric, \
-    RCNNL1LossMetric
-from gluoncv.model_zoo.rcnn.faster_rcnn.data_parallel import ForwardBackwardTask
 
 try:
     import horovod.mxnet as hvd
@@ -320,249 +308,6 @@ def get_dataset(dataset, args):
     return train_dataset, val_dataset, val_metric
 
 
-def get_dataloader(net, train_dataset, val_dataset, train_transform, val_transform, batch_size,
-                   num_shards, args):
-    """Get dataloader."""
-    train_bfn = FasterRCNNTrainBatchify(net, num_shards)
-    if hasattr(train_dataset, 'get_im_aspect_ratio'):
-        im_aspect_ratio = train_dataset.get_im_aspect_ratio()
-    else:
-        im_aspect_ratio = [1.] * len(train_dataset)
-    train_sampler = \
-        gcv.nn.sampler.SplitSortedBucketSampler(im_aspect_ratio, batch_size,
-                                                num_parts=hvd.size() if args.horovod else 1,
-                                                part_index=hvd.rank() if args.horovod else 0,
-                                                shuffle=True)
-    train_loader = mx.gluon.data.DataLoader(train_dataset.transform(
-        train_transform(net.short, net.max_size, net, ashape=net.ashape, multi_stage=args.use_fpn)),
-        batch_sampler=train_sampler, batchify_fn=train_bfn, num_workers=args.num_workers)
-    val_bfn = Tuple(*[Append() for _ in range(3)])
-    short = net.short[-1] if isinstance(net.short, (tuple, list)) else net.short
-    # validation use 1 sample per device
-    val_loader = mx.gluon.data.DataLoader(
-        val_dataset.transform(val_transform(short, net.max_size)), num_shards, False,
-        batchify_fn=val_bfn, last_batch='keep', num_workers=args.num_workers)
-    return train_loader, val_loader
-
-
-def save_params(net, logger, best_map, current_map, epoch, save_interval, prefix):
-    current_map = float(current_map)
-    if current_map > best_map[0]:
-        logger.info('[Epoch {}] mAP {} higher than current best {} saving to {}'.format(
-            epoch, current_map, best_map, '{:s}_best.params'.format(prefix)))
-        best_map[0] = current_map
-        net.save_parameters('{:s}_best.params'.format(prefix))
-        with open(prefix + '_best_map.log', 'a') as f:
-            f.write('{:04d}:\t{:.4f}\n'.format(epoch, current_map))
-    if save_interval and (epoch + 1) % save_interval == 0:
-        logger.info('[Epoch {}] Saving parameters to {}'.format(
-            epoch, '{:s}_{:04d}_{:.4f}.params'.format(prefix, epoch, current_map)))
-        net.save_parameters('{:s}_{:04d}_{:.4f}.params'.format(prefix, epoch, current_map))
-
-
-def split_and_load(batch, ctx_list):
-    """Split data to 1 batch each device."""
-    new_batch = []
-    for i, data in enumerate(batch):
-        if isinstance(data, (list, tuple)):
-            new_data = [x.as_in_context(ctx) for x, ctx in zip(data, ctx_list)]
-        else:
-            new_data = [data.as_in_context(ctx_list[0])]
-        new_batch.append(new_data)
-    return new_batch
-
-
-def validate(net, val_data, ctx, eval_metric, args):
-    """Test on validation dataset."""
-    clipper = gcv.nn.bbox.BBoxClipToImage()
-    eval_metric.reset()
-    if not args.disable_hybridization:
-        # input format is differnet than training, thus rehybridization is needed.
-        net.hybridize(static_alloc=args.static_alloc)
-    for batch in val_data:
-        batch = split_and_load(batch, ctx_list=ctx)
-        det_bboxes = []
-        det_ids = []
-        det_scores = []
-        gt_bboxes = []
-        gt_ids = []
-        gt_difficults = []
-        for x, y, im_scale in zip(*batch):
-            # get prediction results
-            ids, scores, bboxes = net(x)
-            det_ids.append(ids)
-            det_scores.append(scores)
-            # clip to image size
-            det_bboxes.append(clipper(bboxes, x))
-            # rescale to original resolution
-            im_scale = im_scale.reshape((-1)).asscalar()
-            det_bboxes[-1] *= im_scale
-            # split ground truths
-            gt_ids.append(y.slice_axis(axis=-1, begin=4, end=5))
-            gt_bboxes.append(y.slice_axis(axis=-1, begin=0, end=4))
-            gt_bboxes[-1] *= im_scale
-            gt_difficults.append(y.slice_axis(axis=-1, begin=5, end=6) if y.shape[-1] > 5 else None)
-
-        # update metric
-        for det_bbox, det_id, det_score, gt_bbox, gt_id, gt_diff in zip(det_bboxes, det_ids,
-                                                                        det_scores, gt_bboxes,
-                                                                        gt_ids, gt_difficults):
-            eval_metric.update(det_bbox, det_id, det_score, gt_bbox, gt_id, gt_diff)
-    return eval_metric.get()
-
-
-def get_lr_at_iter(alpha, lr_warmup_factor=1. / 3.):
-    return lr_warmup_factor * (1 - alpha) + alpha
-
-
-def train(net, train_data, val_data, eval_metric, batch_size, ctx, args):
-    """Training pipeline"""
-    args.kv_store = 'device' if (args.amp and 'nccl' in args.kv_store) else args.kv_store
-    kv = mx.kvstore.create(args.kv_store)
-    net.collect_params().setattr('grad_req', 'null')
-    net.collect_train_params().setattr('grad_req', 'write')
-    optimizer_params = {'learning_rate': args.lr, 'wd': args.wd, 'momentum': args.momentum}
-    if args.amp:
-        optimizer_params['multi_precision'] = True
-    if args.horovod:
-        hvd.broadcast_parameters(net.collect_params(), root_rank=0)
-        trainer = hvd.DistributedTrainer(
-            net.collect_train_params(),  # fix batchnorm, fix first stage, etc...
-            'sgd',
-            optimizer_params)
-    else:
-        trainer = gluon.Trainer(
-            net.collect_train_params(),  # fix batchnorm, fix first stage, etc...
-            'sgd',
-            optimizer_params,
-            update_on_kvstore=(False if args.amp else None), kvstore=kv)
-
-    if args.amp:
-        amp.init_trainer(trainer)
-
-    # lr decay policy
-    lr_decay = float(args.lr_decay)
-    lr_steps = sorted([float(ls) for ls in args.lr_decay_epoch.split(',') if ls.strip()])
-    lr_warmup = float(args.lr_warmup)  # avoid int division
-
-    # TODO(zhreshold) losses?
-    rpn_cls_loss = mx.gluon.loss.SigmoidBinaryCrossEntropyLoss(from_sigmoid=False)
-    rpn_box_loss = mx.gluon.loss.HuberLoss(rho=args.rpn_smoothl1_rho)  # == smoothl1
-    rcnn_cls_loss = mx.gluon.loss.SoftmaxCrossEntropyLoss()
-    rcnn_box_loss = mx.gluon.loss.HuberLoss(rho=args.rcnn_smoothl1_rho)  # == smoothl1
-    metrics = [mx.metric.Loss('RPN_Conf'),
-               mx.metric.Loss('RPN_SmoothL1'),
-               mx.metric.Loss('RCNN_CrossEntropy'),
-               mx.metric.Loss('RCNN_SmoothL1'), ]
-
-    rpn_acc_metric = RPNAccMetric()
-    rpn_bbox_metric = RPNL1LossMetric()
-    rcnn_acc_metric = RCNNAccMetric()
-    rcnn_bbox_metric = RCNNL1LossMetric()
-    metrics2 = [rpn_acc_metric, rpn_bbox_metric, rcnn_acc_metric, rcnn_bbox_metric]
-
-    # set up logger
-    logging.basicConfig()
-    logger = logging.getLogger()
-    logger.setLevel(logging.INFO)
-    log_file_path = args.save_prefix + '_train.log'
-    log_dir = os.path.dirname(log_file_path)
-    if log_dir and not os.path.exists(log_dir):
-        os.makedirs(log_dir)
-    fh = logging.FileHandler(log_file_path)
-    logger.addHandler(fh)
-    if args.custom_model:
-        logger.info('Custom model enabled. Expert Only!! Currently non-FPN model is not supported!!'
-                    ' Default setting is for MS-COCO.')
-    logger.info(args)
-
-    if args.verbose:
-        logger.info('Trainable parameters:')
-        logger.info(net.collect_train_params().keys())
-    logger.info('Start training from [Epoch {}]'.format(args.start_epoch))
-    best_map = [0]
-    for epoch in range(args.start_epoch, args.epochs):
-        rcnn_task = ForwardBackwardTask(net, trainer, rpn_cls_loss, rpn_box_loss, rcnn_cls_loss,
-                                        rcnn_box_loss, mix_ratio=1.0, amp_enabled=args.amp)
-        executor = Parallel(args.executor_threads, rcnn_task) if not args.horovod else None
-        mix_ratio = 1.0
-        if not args.disable_hybridization:
-            net.hybridize(static_alloc=args.static_alloc)
-        if args.mixup:
-            # TODO(zhreshold) only support evenly mixup now, target generator needs to be modified otherwise
-            train_data._dataset._data.set_mixup(np.random.uniform, 0.5, 0.5)
-            mix_ratio = 0.5
-            if epoch >= args.epochs - args.no_mixup_epochs:
-                train_data._dataset._data.set_mixup(None)
-                mix_ratio = 1.0
-        while lr_steps and epoch >= lr_steps[0]:
-            new_lr = trainer.learning_rate * lr_decay
-            lr_steps.pop(0)
-            trainer.set_learning_rate(new_lr)
-            logger.info("[Epoch {}] Set learning rate to {}".format(epoch, new_lr))
-        for metric in metrics:
-            metric.reset()
-        tic = time.time()
-        btic = time.time()
-        base_lr = trainer.learning_rate
-        rcnn_task.mix_ratio = mix_ratio
-        for i, batch in enumerate(train_data):
-            if epoch == 0 and i <= lr_warmup:
-                # adjust based on real percentage
-                new_lr = base_lr * get_lr_at_iter(i / lr_warmup, args.lr_warmup_factor)
-                if new_lr != trainer.learning_rate:
-                    if i % args.log_interval == 0:
-                        logger.info(
-                            '[Epoch 0 Iteration {}] Set learning rate to {}'.format(i, new_lr))
-                    trainer.set_learning_rate(new_lr)
-            batch = split_and_load(batch, ctx_list=ctx)
-            metric_losses = [[] for _ in metrics]
-            add_losses = [[] for _ in metrics2]
-            if executor is not None:
-                for data in zip(*batch):
-                    executor.put(data)
-            for j in range(len(ctx)):
-                if executor is not None:
-                    result = executor.get()
-                else:
-                    result = rcnn_task.forward_backward(list(zip(*batch))[0])
-                if (not args.horovod) or hvd.rank() == 0:
-                    for k in range(len(metric_losses)):
-                        metric_losses[k].append(result[k])
-                    for k in range(len(add_losses)):
-                        add_losses[k].append(result[len(metric_losses) + k])
-            for metric, record in zip(metrics, metric_losses):
-                metric.update(0, record)
-            for metric, records in zip(metrics2, add_losses):
-                for pred in records:
-                    metric.update(pred[0], pred[1])
-            trainer.step(batch_size)
-
-            # update metrics
-            if (not args.horovod or hvd.rank() == 0) and args.log_interval \
-                    and not (i + 1) % args.log_interval:
-                msg = ','.join(
-                    ['{}={:.3f}'.format(*metric.get()) for metric in metrics + metrics2])
-                logger.info('[Epoch {}][Batch {}], Speed: {:.3f} samples/sec, {}'.format(
-                    epoch, i, args.log_interval * args.batch_size / (time.time() - btic), msg))
-                btic = time.time()
-
-        if (not args.horovod) or hvd.rank() == 0:
-            msg = ','.join(['{}={:.3f}'.format(*metric.get()) for metric in metrics])
-            logger.info('[Epoch {}] Training cost: {:.3f}, {}'.format(
-                epoch, (time.time() - tic), msg))
-            if not (epoch + 1) % args.val_interval:
-                # consider reduce the frequency of validation to save time
-                map_name, mean_ap = validate(net, val_data, ctx, eval_metric, args)
-                val_msg = '\n'.join(['{}={}'.format(k, v) for k, v in zip(map_name, mean_ap)])
-                logger.info('[Epoch {}] Validation: \n{}'.format(epoch, val_msg))
-                current_map = float(mean_ap[-1])
-            else:
-                current_map = 0.
-            save_params(net, logger, best_map, current_map, epoch, args.save_interval,
-                        args.save_prefix)
-
-
 if __name__ == '__main__':
     import sys
 
@@ -574,94 +319,10 @@ if __name__ == '__main__':
     if args.amp:
         amp.init()
 
-    # training contexts
-    if args.horovod:
-        ctx = [mx.gpu(hvd.local_rank())]
-    else:
-        ctx = [mx.gpu(int(i)) for i in args.gpus.split(',') if i.strip()]
-        ctx = ctx if ctx else [mx.cpu()]
-
     # training data
     train_dataset, val_dataset, eval_metric = get_dataset(args.dataset, args)
 
-    # network
-    kwargs = {}
-    module_list = []
-    if args.use_fpn:
-        module_list.append('fpn')
-    if args.norm_layer is not None:
-        module_list.append(args.norm_layer)
-        if args.norm_layer == 'syncbn':
-            kwargs['num_devices'] = len(ctx)
-
-    num_gpus = hvd.size() if args.horovod else len(ctx)
-    net_name = '_'.join(('faster_rcnn', *module_list, args.network, args.dataset))
-    if args.custom_model:
-        args.use_fpn = True
-        net_name = '_'.join(('custom_faster_rcnn_fpn', args.network, args.dataset))
-        if args.norm_layer == 'syncbn':
-            norm_layer = gluon.contrib.nn.SyncBatchNorm
-            norm_kwargs = {'num_devices': len(ctx)}
-            sym_norm_layer = mx.sym.contrib.SyncBatchNorm
-            sym_norm_kwargs = {'ndev': len(ctx)}
-        elif args.norm_layer == 'gn':
-            norm_layer = gluon.nn.GroupNorm
-            norm_kwargs = {'groups': 8}
-            sym_norm_layer = mx.sym.GroupNorm
-            sym_norm_kwargs = {'groups': 8}
-        else:
-            norm_layer = gluon.nn.BatchNorm
-            norm_kwargs = None
-            sym_norm_layer = None
-            sym_norm_kwargs = None
-        classes = train_dataset.CLASSES
-        net = get_model('custom_faster_rcnn_fpn', classes=classes, transfer=None,
-                        dataset=args.dataset, pretrained_base=not args.no_pretrained_base,
-                        base_network_name=args.network, norm_layer=norm_layer,
-                        norm_kwargs=norm_kwargs, sym_norm_kwargs=sym_norm_kwargs,
-                        num_fpn_filters=args.num_fpn_filters,
-                        num_box_head_conv=args.num_box_head_conv,
-                        num_box_head_conv_filters=args.num_box_head_conv_filters,
-                        num_box_head_dense_filters=args.num_box_head_dense_filters,
-                        short=args.image_short, max_size=args.image_max_size, min_stage=2,
-                        max_stage=6, nms_thresh=args.nms_thresh, nms_topk=args.nms_topk,
-                        post_nms=args.post_nms, roi_mode=args.roi_mode, roi_size=args.roi_size,
-                        strides=args.strides, clip=args.clip, rpn_channel=args.rpn_channel,
-                        base_size=args.anchor_base_size, scales=args.anchor_scales,
-                        ratios=args.anchor_aspect_ratio, alloc_size=args.anchor_alloc_size,
-                        rpn_nms_thresh=args.rpn_nms_thresh,
-                        rpn_train_pre_nms=args.rpn_train_pre_nms,
-                        rpn_train_post_nms=args.rpn_train_post_nms,
-                        rpn_test_pre_nms=args.rpn_test_pre_nms,
-                        rpn_test_post_nms=args.rpn_test_post_nms, rpn_min_size=args.rpn_min_size,
-                        per_device_batch_size=args.batch_size // num_gpus,
-                        num_sample=args.rcnn_num_samples, pos_iou_thresh=args.rcnn_pos_iou_thresh,
-                        pos_ratio=args.rcnn_pos_ratio, max_num_gt=args.max_num_gt)
-    else:
-        net = get_model(net_name, pretrained_base=True,
-                        per_device_batch_size=args.batch_size // num_gpus, **kwargs)
-    args.save_prefix += net_name
-    if args.resume.strip():
-        net.load_parameters(args.resume.strip())
-    else:
-        for param in net.collect_params().values():
-            if param._data is not None:
-                continue
-            param.initialize()
-    net.collect_params().reset_ctx(ctx)
-
-    if args.amp:
-        # Cast both weights and gradients to 'float16'
-        net.cast('float16')
-        # These layers don't support type 'float16'
-        net.collect_params('.*batchnorm.*').setattr('dtype', 'float32')
-        net.collect_params('.*normalizedperclassboxcenterencoder.*').setattr('dtype', 'float32')
-
-    # dataloader
-    batch_size = args.batch_size // num_gpus if args.horovod else args.batch_size
-    train_data, val_data = get_dataloader(
-        net, train_dataset, val_dataset, FasterRCNNDefaultTrainTransform,
-        FasterRCNNDefaultValTransform, batch_size, len(ctx), args)
+    frcnn_estimator = FasterRCNNEstimator(args, eval_metric)
 
     # training
-    train(net, train_data, val_data, eval_metric, batch_size, ctx, args)
+    frcnn_estimator.fit(train_dataset, val_dataset)
